Add initial Dockerfile
It would be nicer to use a smaller base image such as alpine. However,
Go will link to glibc when building the net package - alpine comes with
musl libc. We could disable cgo and ask the compiler to use the Go
implementations of the net dependencies, but we may be asking for trouble.