Send /data and /latest requests directly to babbage
Instead of letting /data and /latest requests into the allRoutesHandler, send them directly to Babbage. The allRoutesHandler make a request to zebedee to try and determine the page type. Requests to /data will not be recognised by Zebedee. Requests to /latest are only supported by the page types bulletin, article, and compendium. Only routing dataset types are rerouted in the allRoutesHandler, it's more optimal to send /latest requests directly to Babbage.